refactor(arch): move call_context into services layer
The request-scoped CallContext ContextVar is an application-layer
propagation primitive (set by services, read by adapters), not an
adapter — it has no infrastructure surface. Hosting it under
qfa.adapters forced services -> adapters imports that violated the
hexagonal layer contract.

Move src/qfa/adapters/call_context.py to src/qfa/services/call_context.py
and update the five import sites; relocate the matching test module
from tests/adapters/ to tests/services/. Repair an orphan test
(test_analyze_enters_call_scope_with_analyze_operation) that had been
constructing Orchestrator directly without anonymizer; route it through
the existing _make_orchestrator factory like every other test in the
file.

Extend the api -> adapters composition-root allowlist in
pyproject.toml to cover qfa.adapters.tracking_llm and qfa.adapters.db,
which the FastAPI lifespan now wires. import-linter contracts: 3 kept,
0 broken; full suite: 180 passed.
